+++ src/main/kotlin/org/ornet/pat/localization/db/v2/LocalizedTextModel.kt
@@ -18,4 +18,3 @@ data class LocalizedText(
     val lines: Int,
     val version: Long,
 )
-